+++ 0cchext/inc/engextcpp.cpp
@@ -222,7 +222,7 @@ ExtCommandDesc::ExtCommandDesc(__in PCSTR Name,
 
     if (strlen(Name) > s_LongestCommandName)
     {
-        s_LongestCommandName = strlen(Name);
+        s_LongestCommandName = (ULONG)strlen(Name);
     }
 }
 
@@ -1200,7 +1200,7 @@ ExtExtension::OutWrapStr(__in PCSTR String)
 {
     if (m_TestWrap)
     {
-        m_TestWrapChars += strlen(String);
+        m_TestWrapChars += (ULONG)strlen(String);
         return;
     }
     
@@ -1303,7 +1303,7 @@ ExtExtension::CopyCircleString(__in PCSTR Str)
     PSTR Buf;
     ULONG Chars;
     
-    Chars = strlen(Str) + 1;
+    Chars = (ULONG)strlen(Str) + 1;
     Buf = RequestCircleString(Chars);
     memcpy(Buf, Str, Chars * sizeof(*Str));
     return Buf;
@@ -1344,7 +1344,7 @@ ExtExtension::AppendBufferString(__in PCSTR Str)
 {
     ULONG Chars;
     
-    Chars = strlen(Str) + 1;
+    Chars = (ULONG)strlen(Str) + 1;
     if (Chars > m_AppendBufferChars ||
         (ULONG_PTR)(m_AppendAt - m_AppendBuffer) > m_AppendBufferChars - Chars)
     {
@@ -1785,7 +1785,7 @@ ExtExtension::CallDebuggeeBase(__in PCSTR CommandString,
     ExtCaptureOutputA IgnoreOut;
 
     Cmd.Copy(".call ", 6);
-    Cmd.Append(CommandString, strlen(CommandString) + 1);
+    Cmd.Append(CommandString, (ULONG)strlen(CommandString) + 1);
 
     if (FAILED(Status = m_Control->
                Execute(DEBUG_OUTCTL_IGNORE,
@@ -2611,7 +2611,7 @@ ExtExtension::HandleKnownStruct(__in PDEBUG_CLIENT Client,
         Status = S_OK;
         while (Struct && Struct->TypeName)
         {
-            ULONG Chars = strlen(Struct->TypeName) + 1;
+            ULONG Chars = (ULONG)strlen(Struct->TypeName) + 1;
             CharsNeeded += Chars;
             
             if (Status != S_OK || *BufferChars < Chars)
@@ -2711,7 +2711,7 @@ ExtExtension::HandleQueryValueNames(__in PDEBUG_CLIENT Client,
     Status = S_OK;
     while (ExtVal && ExtVal->ValueName)
     {
-        ULONG Chars = wcslen(ExtVal->ValueName) + 1;
+        ULONG Chars = (ULONG)wcslen(ExtVal->ValueName) + 1;
         CharsNeeded += Chars;
             
         if (Status != S_OK || BufferChars < Chars)
@@ -3612,7 +3612,7 @@ ExtExtension::HelpAll(void)
     ExtCommandDesc* Desc = m_Commands;
     while (Desc)
     {
-        ULONG NameLen = strlen(Desc->m_Name);
+        ULONG NameLen = (ULONG)strlen(Desc->m_Name);
         OutWrap("  !%s%*c- ",
                 Desc->m_Name,
                 m_LongestCommandName - NameLen + 1, ' ');
@@ -4352,7 +4352,7 @@ ExtRemoteTyped::ErtIoctl(__in PCSTR Message,
         StrBufferChars * sizeof(*StrBuffer);
     if (InStr)
     {
-        ExtDataBytes += (strlen(InStr) + 1) * sizeof(*InStr);
+        ExtDataBytes += ((ULONG)strlen(InStr) + 1) * sizeof(*InStr);
     }
 
     ExtData = (EXT_TYPED_DATA*)DataHolder.Get(ExtDataBytes);
